+++ stdlib/Dates/src/periods.jl
@@ -197,6 +197,16 @@ struct CompoundPeriod <: AbstractTime
 end
 
 """
+    Dates.periods(::CompoundPeriod) -> Vector{Period}
+
+Return the `Vector` of `Period`s that comprise the given `CompoundPeriod`.
+
+!!! compat "Julia 1.7"
+    This function requires Julia 1.7 or later.
+"""
+periods(x::CompoundPeriod) = x.periods
+
+"""
     CompoundPeriod(periods) -> CompoundPeriod
 
 Construct a `CompoundPeriod` from a `Vector` of `Period`s. All `Period`s of the same type